fix: quote leading-dash gh upload arg; split oversized test file
Codex finding on commit e5b9c40: `gh release upload`'s own arg parser
treats a leading '-' in a file argument as a FLAG, not part of the
filename (confirmed against real gh 2.96.0: `gh release upload dummy
--nightly.tar.gz` exits "unknown flag" rather than uploading a file by
that name). A resolved asset name starting with '-' packages and stages
just fine in actions/stage-baseline -- '-' is a perfectly legal leading
filename character on every real filesystem -- but this step's
first-time-publish `gh release upload "$RELEASE_TAG" "$ASSET_NAME"` call
would otherwise always fail. Fixed by passing "./$ASSET_NAME" instead of
a bare "$ASSET_NAME". New regression test confirmed to fail against the
pre-fix code and pass with the fix.

This latest round of TestUploadReleaseAsset* additions pushed
tests/test_publish_baseline_workflows.py to 2075 lines, past the
AI-readiness 2000-line hard cap. Split the file: the 9 (now 10)
TestUploadReleaseAsset* classes -- covering the "Upload release asset"
step's several distinct safe-retry/immutability rejection shapes --
move to a new sibling tests/test_publish_baseline_upload_step.py, with
their own small, deliberately independent copy of the shared
_load/_steps/_bash_executable/_WINDOWS_PYTHON3_SKIP helpers (kept
non-cross-module so each file stays independently collectible, matching
this repo's existing test-suite convention of no cross-test-module
imports). All 61 tests (51 + 10) preserved; nothing renamed or dropped.
